Clean up utils/ sprawl: delete dead code, remove duplicates, flatten structure
- Delete dead code: logging.py (unused re-export wrapper), r_conversion.py (0 importers)
- Delete duplicate date_utils.py (top-level 70-line version superseded by common/date_utils.py)
- Merge USAspendingCache into APICache usage (identical behavior with default_cache_type="contracts")
- Flatten common/ subdirectory: move date_utils, logging_config, path_utils up to utils/
- Consolidate enrichment_checkpoints.py, enrichment_freshness.py, enrichment_metrics.py into enrichment/ package
- Clean up __init__.py: use consistent relative imports, reorder by category

Reduces utils/ from 48 files to 44, removes ~1,570 lines of duplicate/dead code.
Pre-existing broken imports in examples/ and scripts/ (from prior monorepo split) are unchanged.
