+++ sbir_etl/enrichers/usaspending/client.py
@@ -62,7 +62,7 @@ class USAspendingAPIClient:
         self.state_file = Path(
             self.config.get("state_file", "data/state/enrichment_refresh_state.json")
         )
-        from sbir_etl.utils.common.path_utils import ensure_parent_dir
+        from sbir_etl.utils.path_utils import ensure_parent_dir
 
         ensure_parent_dir(self.state_file)
 
@@ -533,7 +533,7 @@ class USAspendingAPIClient:
             state: State dictionary to save
         """
         try:
-            from sbir_etl.utils.common.path_utils import ensure_parent_dir
+            from sbir_etl.utils.path_utils import ensure_parent_dir
 
             ensure_parent_dir(self.state_file)
             with open(self.state_file, "w") as f:
