+++ sbir_etl/models/award.py
@@ -810,7 +810,7 @@ class RawAward(BaseModel):
         data = self.model_dump()
 
         # Parse dates using centralized utility
-        from sbir_etl.utils.common.date_utils import parse_date
+        from sbir_etl.utils.date_utils import parse_date
 
         date_fields = [
             "award_date",
