+++ sbir_etl/models/contract_models.py
@@ -98,7 +98,7 @@ class ContractPeriod(BaseModel):
     )
     @classmethod
     def parse_dates(cls, v):
-        from sbir_etl.utils.common.date_utils import parse_date
+        from sbir_etl.utils.date_utils import parse_date
 
         result = parse_date(v, strict=True)
         if result is None:
