+++ sbir_etl/transformers/patent_transformer.py
@@ -202,7 +202,7 @@ class PatentAssignmentTransformer:
     @staticmethod
     def _parse_date(value: Any | None) -> date | None:
         """Parse date using centralized utility."""
-        from sbir_etl.utils.common.date_utils import parse_date
+        from sbir_etl.utils.date_utils import parse_date
 
         return parse_date(value)
