+++ sbir_etl/utils/__init__.py
@@ -4,13 +4,15 @@
 from .cache.base_cache import BaseDataFrameCache
 
 # Date utilities
-from sbir_etl.utils.common.date_utils import (
+from .date_utils import (
     format_date_iso,
     parse_date,
     parse_date_safe,
     validate_date_range,
 )
-from sbir_etl.utils.common.logging_config import (
+
+# Logging utilities
+from .logging_config import (
     LogContext,
     configure_logging_from_config,
     log_debug,
@@ -23,7 +25,7 @@ from sbir_etl.utils.common.logging_config import (
 )
 
 # Path utilities
-from sbir_etl.utils.common.path_utils import (
+from .path_utils import (
     ensure_dir,
     ensure_parent_dir,
     ensure_path_exists,
@@ -33,17 +35,19 @@ from sbir_etl.utils.common.path_utils import (
 )
 
 # Chunking utilities
-from sbir_etl.utils.data.chunking import (
+from .data.chunking import (
     ChunkIterator,
     batch_process,
     chunk_dataframe,
     chunk_generator,
     chunk_iterable,
 )
-from sbir_etl.utils.data.duckdb_client import DuckDBClient, get_duckdb_client
+
+# DuckDB utilities
+from .data.duckdb_client import DuckDBClient, get_duckdb_client
 
 # File I/O utilities
-from sbir_etl.utils.data.file_io import (
+from .data.file_io import (
     read_parquet_or_ndjson,
     save_dataframe_parquet,
     write_json,
@@ -51,9 +55,8 @@ from sbir_etl.utils.data.file_io import (
     write_ndjson,
 )
 
+# Asset column utilities
 from .asset_column_helper import AssetColumnHelper
-
-# Column discovery utilities
 from .column_finder import ColumnFinder
 
 # Configuration utilities
@@ -61,13 +64,22 @@ from .config_accessor import ConfigAccessor
 
 # Error handling utilities
 from .error_handling import handle_asset_error, log_and_raise, retry_with_backoff, safe_execute
+
+# Metrics
 from .metrics import MetricsCollector, PipelineMetrics
+
+# Text normalization
 from .text_normalization import normalize_company_name, normalize_name, normalize_recipient_name
 
 
 __all__ = [
     # Cache
     "BaseDataFrameCache",
+    # Date utilities
+    "parse_date",
+    "parse_date_safe",
+    "format_date_iso",
+    "validate_date_range",
     # Logging
     "LogContext",
     "configure_logging_from_config",
@@ -78,21 +90,16 @@ __all__ = [
     "log_warning",
     "log_with_context",
     "setup_logging",
+    # Path utilities
+    "ensure_dir",
+    "ensure_parent_dir",
+    "ensure_path_exists",
+    "normalize_path_list",
+    "resolve_path",
+    "safe_path_join",
     # DuckDB
     "DuckDBClient",
     "get_duckdb_client",
-    # Metrics
-    "MetricsCollector",
-    "PipelineMetrics",
-    # Text normalization
-    "normalize_name",
-    "normalize_company_name",
-    "normalize_recipient_name",
-    # Column discovery
-    "ColumnFinder",
-    "AssetColumnHelper",
-    # Configuration
-    "ConfigAccessor",
     # File I/O
     "save_dataframe_parquet",
     "write_json",
@@ -105,21 +112,21 @@ __all__ = [
     "chunk_generator",
     "ChunkIterator",
     "batch_process",
+    # Column discovery
+    "ColumnFinder",
+    "AssetColumnHelper",
+    # Configuration
+    "ConfigAccessor",
     # Error handling
     "handle_asset_error",
     "log_and_raise",
     "retry_with_backoff",
     "safe_execute",
-    # Date utilities
-    "parse_date",
-    "parse_date_safe",
-    "format_date_iso",
-    "validate_date_range",
-    # Path utilities
-    "ensure_dir",
-    "ensure_parent_dir",
-    "ensure_path_exists",
-    "normalize_path_list",
-    "resolve_path",
-    "safe_path_join",
+    # Metrics
+    "MetricsCollector",
+    "PipelineMetrics",
+    # Text normalization
+    "normalize_name",
+    "normalize_company_name",
+    "normalize_recipient_name",
 ]
