+++ sbir_etl/utils/enrichment/checkpoints.py
@@ -76,7 +76,7 @@ class CheckpointStore:
         if parquet_path is None:
             parquet_path = Path("data/state/enrichment_checkpoints.parquet")
         self.parquet_path = Path(parquet_path)
-        from sbir_etl.utils.common.path_utils import ensure_parent_dir
+        from sbir_etl.utils.path_utils import ensure_parent_dir
 
         ensure_parent_dir(self.parquet_path)
