+++ scripts/backfill_enrichment_freshness.py
@@ -23,7 +23,7 @@ import pandas as pd
 from loguru import logger
 
 from sbir_etl.models.enrichment import EnrichmentFreshnessRecord, EnrichmentStatus
-from sbir_etl.utils.enrichment_freshness import FreshnessStore
+from sbir_etl.utils.enrichment.freshness import FreshnessStore
 
 
 def compute_enrichment_hash(enriched_data: dict[str, Any]) -> str:
