+++ tests/unit/utils/test_enrichment_checkpoints.py
@@ -21,7 +21,7 @@ pytestmark = pytest.mark.fast
 import pandas as pd
 import pytest
 
-from sbir_etl.utils.enrichment_checkpoints import CheckpointStore, EnrichmentCheckpoint
+from sbir_etl.utils.enrichment.checkpoints import CheckpointStore, EnrichmentCheckpoint
 
 
 pytestmark = pytest.mark.fast
